refactor: centralize method naming code in a single function
