+++ paper_experiments/figures/generate_heterogeneity_plots_utils.py
@@ -11,54 +11,10 @@ from fedpydeseq2_datasets.utils import get_experiment_id
 from loguru import logger
 from matplotlib import pyplot as plt
 
-from paper_experiments.figures.generate_cross_tables_utils import NAME_MAPPING
-from paper_experiments.figures.generate_cross_tables_utils import (
-    get_padj_lfc_from_method,
-)
+from paper_experiments.figures.utils import get_de_genes
+from paper_experiments.figures.utils import get_padj_lfc_from_method
+from paper_experiments.figures.utils import process_method_name
 from paper_experiments.utils.constants import MetaAnalysisParameter
-
-
-def get_de_genes(
-    method_padj: pd.Series,
-    method_lfc: pd.Series,
-    padj_threshold: float | None,
-    log2fc_threshold: float | None,
-) -> pd.Index:
-    """
-    Get the differentially expressed genes.
-
-    We define the differentially expressed genes as the genes with an adjusted p-value
-    below a certain threshold and an absolute log fold change above a certain threshold.
-
-    Parameters
-    ----------
-    method_padj : pd.Series
-        The adjusted p-values, indexed by gene names.
-
-    method_lfc : pd.Series
-        The log fold changes, indexed by gene names, *in natural scale*.
-
-    padj_threshold : float or None
-        The adjusted p-value threshold.
-
-    log2fc_threshold : float or None
-        The log2 fold change threshold.
-
-    Returns
-    -------
-    pd.Index
-        The differentially expressed genes.
-
-    """
-    # Initialize a boolean series to True
-    condition = pd.Series(True, index=method_padj.index)
-    if padj_threshold is not None:
-        condition &= method_padj < padj_threshold
-
-    if log2fc_threshold is not None:
-        condition &= np.abs(method_lfc) > np.log(2) * log2fc_threshold
-    method_diff_genes = method_padj[condition].index
-    return method_diff_genes
 
 
 def sensitivity(
@@ -537,7 +493,7 @@ def build_heterogeneity_grid_plot(
                 else:
                     assert isinstance(method_test_padj, pd.Series)
                     assert isinstance(method_test_lfc, pd.Series)
-                    method_test_str = NAME_MAPPING[method_test]
+                    method_test_str = process_method_name(method_test)
                     if heterogeneity_id == 0:
                         methods_test_padj_lfc[method_test_str] = []
                     methods_test_padj_lfc[method_test_str].append(
@@ -585,7 +541,7 @@ def build_heterogeneity_grid_plot(
                 for k, score in enumerate(scores_list):  # TODO check silent variable
                     lines.append(
                         {
-                            "method_test_name": process_method_test_name(method_test),
+                            "method_test_name": process_method_name(method_test),
                             # Here we invert the heterogeneity level
                             "heterogeneity level": 1.0 - heterogeneity_method_params[k],
                             "score": score,
@@ -804,7 +760,7 @@ def build_test_vs_ref_heterogeneity_plot(
             else:
                 assert isinstance(method_test_padj, pd.Series)
                 assert isinstance(method_test_lfc, pd.Series)
-                method_test_str = NAME_MAPPING[method_test]
+                method_test_str = process_method_name(method_test)
                 if heterogeneity_id == 0:
                     methods_test_padj_lfc[method_test_str] = []
                 methods_test_padj_lfc[method_test_str].append(
@@ -865,49 +821,6 @@ def build_test_vs_ref_heterogeneity_plot(
     )
 
 
-def process_method_test_name(method_test_name: str) -> str:
-    """
-    Process the method test name for plots.
-
-    Parameters
-    ----------
-    method_test_name : str
-        The method test name.
-
-    Returns
-    -------
-    str
-        The processed method test name.
-
-    """
-    if method_test_name.startswith("FedPyDESeq2"):
-        return "\n".join(method_test_name.split(", "))
-    elif method_test_name.startswith("Meta-analysis"):
-        # Split the method name
-        method_test_name_split = method_test_name.split(", ")
-        # Get the meta-analysis type
-        meta_analysis_type = method_test_name_split[1]
-        # Get the method combination
-        method_combination = method_test_name_split[2]
-        # Get the pvalue combination
-        pvalue_combination = method_test_name_split[3]
-        if meta_analysis_type == "random_effect":
-            if method_combination == "dl":
-                return "Random effect\n(DerSimonian-Laird)"
-            else:
-                return f"Random effect\n({method_combination})"
-        elif meta_analysis_type == "fixed_effect":
-            return "Fixed effect"
-        elif meta_analysis_type == "pvalue_combination":
-            return f"P-value combination\n({pvalue_combination})"
-        else:
-            raise ValueError(f"Unknown meta-analysis type: {meta_analysis_type}")
-    elif method_test_name.startswith("PyDESeq2"):
-        return "\n".join(method_test_name.split(", "))
-    else:
-        raise ValueError(f"Unknown method test name: {method_test_name}")
-
-
 def make_heterogeneity_plot(
     scores: dict[str, list[float]],
     heterogeneity_method_params: list[float],
@@ -947,7 +860,7 @@ def make_heterogeneity_plot(
         for i, score in enumerate(scores_list):
             lines.append(
                 {
-                    "method_test_name": process_method_test_name(method_test),
+                    "method_test_name": process_method_name(method_test),
                     # Here we invert the heterogeneity level
                     "heterogeneity level": heterogeneity_method_params_names[i]
                     if heterogeneity_method_params_names is not None
